+++ tests/test_sanity.py
@@ -13,12 +13,14 @@ from subprocess import PIPE, STDOUT
 
 from runner import RunnerCore, path_from_root, env_modify, test_file
 from runner import create_file, ensure_dir, make_executable, with_env_modify
+from runner import parameterized
 from tools.config import EM_CONFIG
 from tools.shared import EMCC
 from tools.shared import CANONICAL_TEMP_DIR
 from tools.shared import try_delete, config
 from tools.shared import EXPECTED_LLVM_VERSION, Cache
 from tools import shared, system_libs, utils
+from tools import response_file
 
 SANITY_FILE = shared.Cache.get_path('sanity.txt')
 commands = [[EMCC], [path_from_root('tests', 'runner'), 'blahblah']]
@@ -489,7 +491,41 @@ fi
     # Exactly one child process should have triggered libc build!
     self.assertEqual(num_times_libc_was_built, 1)
 
-  def test_emconfig(self):
+  @parameterized({
+    '': [False],
+    'response_files': [True]
+  })
+  def test_emcc_cache_flag(self, use_response_files):
+    restore_and_set_up()
+
+    cache_dir_name = self.in_dir('emscripten_cache')
+    self.assertFalse(os.path.exists(cache_dir_name))
+    create_file('test.c', r'''
+      #include <stdio.h>
+      int main() {
+        printf("hello, world!\n");
+        return 0;
+      }
+      ''')
+    args = ['--cache', cache_dir_name]
+    if use_response_files:
+      rsp = response_file.create_response_file(args, shared.TEMP_DIR)
+      args = ['@' + rsp]
+
+    self.run_process([EMCC, 'test.c'] + args, stderr=PIPE)
+    if use_response_files:
+      os.remove(rsp)
+
+    # The cache directory must exist after the build
+    self.assertTrue(os.path.exists(cache_dir_name))
+    # The cache directory must contain a sysroot
+    self.assertTrue(os.path.exists(os.path.join(cache_dir_name, 'sysroot')))
+
+  @parameterized({
+    '': [False],
+    'response_files': [True]
+  })
+  def test_emconfig(self, use_response_files):
     restore_and_set_up()
 
     fd, custom_config_filename = tempfile.mkstemp(prefix='.emscripten_config_')
@@ -505,9 +541,17 @@ fi
 
     temp_dir = tempfile.mkdtemp(prefix='emscripten_temp_')
 
+    args = ['--em-config', custom_config_filename]
+    if use_response_files:
+      rsp = response_file.create_response_file(args, shared.TEMP_DIR)
+      args = ['@' + rsp]
+
     with utils.chdir(temp_dir):
-      self.run_process([EMCC, '--em-config', custom_config_filename] + MINIMAL_HELLO_WORLD + ['-O2'])
+      self.run_process([EMCC] + args + MINIMAL_HELLO_WORLD + ['-O2'])
       result = self.run_js('a.out.js')
+
+    if use_response_files:
+      os.remove(rsp)
 
     self.assertContained('hello, world!', result)